+++ src/plugins/empty-message/empty-message.tsx
@@ -9,8 +9,6 @@ export function create<T>(config: Config<T>): GridPlugin<T> {
   return {
     name: "empty-message",
     reducer: (s) => s,
-    footer: {
-      span: (data) => data.length === 0 && config.message
-    }
+    footerSpan: (data) => data.length === 0 && config.message,
   };
 }